+++ src/main/java/palamod/procedures/Chervil0upProcedure.java
@@ -38,9 +38,8 @@ public class Chervil0upProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if (new Random().nextInt(250 + 1) == 1
-				&& (world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == Blocks.FARMLAND) {
-			world.setBlockState(new BlockPos((int) x, (int) y, (int) z), Chervil1Block.block.getDefaultState(), 3);
+		if (new Random().nextInt(250 + 1) == 1 && (world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == Blocks.FARMLAND) {
+			world.setBlockState(new BlockPos(x, y, z), Chervil1Block.block.getDefaultState(), 3);
 		}
 	}
 }
