+++ src/main/java/palamod/procedures/Clear1Procedure.java
@@ -142,7 +142,7 @@ public class Clear1Procedure {
 			}
 		}
 		if (world instanceof World && !world.isRemote()) {
-			((World) world).playSound(null, new BlockPos((int) x, (int) y, (int) z),
+			((World) world).playSound(null, new BlockPos(x, y, z),
 					(net.minecraft.util.SoundEvent) ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("block.anvil.use")),
 					SoundCategory.NEUTRAL, (float) 1, (float) 1);
 		} else {
