+++ src/main/java/palamod/procedures/Eggplant3shearProcedure.java
@@ -52,7 +52,7 @@ public class Eggplant3shearProcedure {
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
 		Entity entity = (Entity) dependencies.get("entity");
 		if (((entity instanceof LivingEntity) ? ((LivingEntity) entity).getHeldItemMainhand() : ItemStack.EMPTY).getItem() == Items.SHEARS) {
-			world.setBlockState(new BlockPos((int) x, (int) y, (int) z), Eggplant1Block.block.getDefaultState(), 3);
+			world.setBlockState(new BlockPos(x, y, z), Eggplant1Block.block.getDefaultState(), 3);
 			if (Math.random() < 0.2) {
 				if (world instanceof World && !world.isRemote()) {
 					ItemEntity entityToSpawn = new ItemEntity((World) world, x, y, z, new ItemStack(EggplantseedItem.block));
