+++ src/main/java/palamod/procedures/Grinder2connectpalahelpProcedure.java
@@ -63,7 +63,7 @@ public class Grinder2connectpalahelpProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -82,7 +82,7 @@ public class Grinder2connectpalahelpProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -100,7 +100,7 @@ public class Grinder2connectpalahelpProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
