+++ src/main/java/palamod/procedures/OpencheckssetuppalahelpProcedure.java
@@ -61,7 +61,7 @@ public class OpencheckssetuppalahelpProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -79,7 +79,7 @@ public class OpencheckssetuppalahelpProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
